+++ DemoApp/ViewController.swift
@@ -1,0 +1,30 @@
+//
+//  ViewController.swift
+//  DemoApp
+//
+//  Created by dan on 07/07/2017.
+//  Copyright © 2017 Hydra. All rights reserved.
+//
+
+import UIKit
+import Hydra
+
+public enum Errors: Error {
+	case any
+}
+
+class ViewController: UIViewController {
+	
+	override func viewDidLoad() {
+		super.viewDidLoad()
+		// Do any additional setup after loading the view, typically from a nib.
+	}
+
+	override func didReceiveMemoryWarning() {
+		super.didReceiveMemoryWarning()
+		// Dispose of any resources that can be recreated.
+	}
+
+
+}
+