+++ Sources/Hydra/Context.swift
@@ -63,5 +63,5 @@ public enum Context {
 			case .custom(let queue): return queue
 		}
 	}
-	
 }
+